+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased](https://github.com/contentful/ContentfulBundle/compare/0.5.0-beta...HEAD)
+## [Unreleased](https://github.com/contentful/ContentfulBundle/compare/0.6.0-beta...HEAD)
+
+## [0.5.0-beta](https://github.com/contentful/ContentfulBundle/tree/0.6.0-beta) (2016-03-03)
 
 ### Added
 * Show metrics about requests in the Web Profiler